LPS-169327 Merge ViewClientExtensionEntryPartDisplayContext into ViewClientExtensionEntryDisplayContext
