+++ modules/apps/client-extension/client-extension-web/src/main/java/com/liferay/client/extension/web/internal/constants/ClientExtensionAdminWebKeys.java
@@ -28,8 +28,4 @@ public class ClientExtensionAdminWebKeys {
 	public static final String VIEW_CLIENT_EXTENSION_ENTRY_DISPLAY_CONTEXT =
 		"VIEW_CLIENT_EXTENSION_ENTRY_DISPLAY_CONTEXT";
 
-	public static final String
-		VIEW_CLIENT_EXTENSION_ENTRY_PART_DISPLAY_CONTEXT =
-			"VIEW_CLIENT_EXTENSION_ENTRY_PART_DISPLAY_CONTEXT";
-
 }
