+++ modules/apps/client-extension/client-extension-web/src/main/resources/META-INF/resources/admin/init.jsp
@@ -31,7 +31,6 @@ page import="com.liferay.client.extension.web.internal.constants.ClientExtension
 page import="com.liferay.client.extension.web.internal.display.context.ClientExtensionAdminDisplayContext" %><%@
 page import="com.liferay.client.extension.web.internal.display.context.EditClientExtensionEntryDisplayContext" %><%@
 page import="com.liferay.client.extension.web.internal.display.context.ViewClientExtensionEntryDisplayContext" %><%@
-page import="com.liferay.client.extension.web.internal.display.context.ViewClientExtensionEntryPartDisplayContext" %><%@
 page import="com.liferay.portal.kernel.service.WorkflowDefinitionLinkLocalServiceUtil" %>
 
 <%@ page import="java.lang.reflect.Method" %>
